requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
pytesseract==0.3.10
pdf2image==1.16.3
python-multipart==0.0.6
pillow==10.1.0
pydantic==2.4.2
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1 